[ENGN-9020] Share the docs page model between generated bundles
The navigation walk, the frontmatter parse and the MDX -> markdown
reduction lived inside scripts/generateLlmsTxt.js, so a second generator
that needs the same page model had only two options: import a script
that runs on require, or copy it. Copying is the worse one — two walks
that can disagree about which pages exist, and two reductions that can
disagree about what a page's markdown is.

They now live in scripts/lib/docsPages.js, which owns the page set
(collectAllPages, including the invariant that the meta-driven walk
reaches every page on disk) and the reduction (mdxToMarkdown).
generateLlmsTxt.js keeps only what is specific to the two bundles: the
site summary, the index and full-text rendering, and its check mode.

Behaviour-preserving: public/llms.txt and public/llms-full.txt
regenerate byte for byte, and --check still passes. The one addition is
mdxToMarkdown's stripLeadingH1 option — llms-full.txt renders its own
title header above each page and so drops the page's H1, which a
standalone per-page file must keep — defaulting to the existing
behaviour.
